requirements.txt
groq
langchain-groq
streamlit
langchain_community
langchain_core
python-dotenv
pypdf
langchain-ollama
faiss-cpu